Revert existent test changes and creating specific tests for span decorators
