+++ instrument-starters/opentracing-spring-cloud-feign-starter/src/test/java/io/opentracing/contrib/spring/cloud/feign/FeignManualWithSpanDecoratorsTest.java
@@ -13,63 +13,31 @@
  */
 package io.opentracing.contrib.spring.cloud.feign;
 
-import static io.opentracing.contrib.spring.cloud.feign.FeignTest.verify;
+import static io.opentracing.contrib.spring.cloud.feign.TestUtils.verifyWithSpanDecorators;
 
-import feign.Client;
-import feign.Feign;
-import feign.RequestLine;
 import io.opentracing.contrib.spring.cloud.feign.FeignManualTest.ManualFeignConfiguration;
 import io.opentracing.contrib.spring.cloud.feign.FeignTest.FeignRibbonLocalConfiguration;
-import io.opentracing.mock.MockTracer;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.cloud.netflix.feign.FeignClient;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 
 /**
- * @author Pavol Loffay
+ * @author Emerson Oliveira
  */
 @SpringBootTest(
     webEnvironment = WebEnvironment.RANDOM_PORT,
     classes = {MockTracingConfiguration.class, TestController.class,
-        ManualFeignConfiguration.class, FeignRibbonLocalConfiguration.class},
+        ManualFeignConfiguration.class, FeignRibbonLocalConfiguration.class,
+        FeignSpanDecoratorConfiguration.class},
     properties = {"opentracing.spring.web.skipPattern=/notTraced"})
 @RunWith(SpringJUnit4ClassRunner.class)
-public class FeignManualTest {
-
-  @Configuration
-  static class ManualFeignConfiguration {
-
-    @Autowired
-    public ManualFeignConfiguration(Client client) {
-      feignInterface = Feign.builder().client(client)
-          .target(FeignInterface.class, "http://localService");
-    }
-  }
-
-  @FeignClient(value = "localService")
-  interface FeignInterface {
-
-    @RequestLine("GET /notTraced")
-    // TODO this has to be added when using spring Boot 1.4.1
-    @RequestMapping(method = RequestMethod.GET, value = "/notTraced")
-    String hello();
-  }
-
-  protected static FeignInterface feignInterface;
-
-  @Autowired
-  protected MockTracer mockTracer;
+public class FeignManualWithSpanDecoratorsTest extends FeignManualTest {
 
   @Test
   public void testTracedRequestDefinedUrl() throws InterruptedException {
     feignInterface.hello();
-    verify(mockTracer);
+    verifyWithSpanDecorators(mockTracer);
   }
 }
